Mitigate Issue 2085
Previously, comment length was calculated by taking the length of the comment array. However, this does not factor in the length of replies. This commit factors that in. However, it still does not fully solve the problem. I suspect this is because `projectInfo.stats.comments.length` does not take into account deleted comments.